Refactor completion provider
Create a shared ReplCommandCompletionProvider between VB and C#.
